@@ -1,43 +1,59 @@
 package outout.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.Environment;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.config.authentication.AuthenticationManagerBeanDefinitionParser;
-import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+import org.springframework.security.web.context.SecurityContextRepository;
 import outout.security.StatelessAuthenticationFilter;
 
-import javax.servlet.Filter;
+import jakarta.servlet.Filter;
+import static org.springframework.security.config.http.SessionCreationPolicy.STATELESS;
 
 @Configuration
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
+@EnableWebSecurity
+public class SecurityConfig {
     @Autowired
     private Environment environment;
 
-    @Override
-    protected void configure(final HttpSecurity http) throws Exception {
-        http.csrf().disable()
-                .headers().frameOptions().disable()
-                .and()
-                .authorizeRequests()
-                .antMatchers("/account/create").permitAll()
-                .antMatchers("/authenticate").permitAll()
-                .antMatchers("/console/**").permitAll()
-                .antMatchers("/swagger-ui.html").permitAll()
-                .antMatchers("/swagger").permitAll()
-                .antMatchers("/v2/api-docs").permitAll()
-                .antMatchers("/webjars/**").permitAll()
-                .antMatchers("/swagger-resources/**").permitAll()
-                .anyRequest().authenticated()
-                .and()
-                .authenticationProvider(new AuthenticationManagerBeanDefinitionParser.NullAuthenticationProvider())
-        .addFilterBefore(statelessAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
+    @Autowired
+    private SecurityContextRepository repository;
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http
+                .authorizeHttpRequests((authz) -> authz
+                        .anyRequest().authenticated()
+                )
+                .sessionManagement(manager -> manager.sessionCreationPolicy(STATELESS))
+                .addFilterBefore(statelessAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class)
+                .csrf(csrf -> csrf.disable());
+        return http.build();
     }
 
+    @Bean
+    public WebSecurityCustomizer webSecurityCustomizer() {
+        return (web) -> web.ignoring().requestMatchers((request) -> {
+            String path = request.getServletPath();
+            return (path.startsWith("/authenticate") ||
+                    path.startsWith("/swagger-ui.html") ||
+                    path.startsWith("/swagger") ||
+                    path.startsWith("/swagger-ui") ||
+                    path.startsWith("/v3/api-docs") ||
+                    path.startsWith("/webjars/") ||
+                    path.startsWith("/swagger-resources/") ||
+                    path.startsWith("/account/create") ||
+                    path.startsWith("/error") );
+        });
+    }
+
+
     private Filter statelessAuthenticationFilter() {
-        return new StatelessAuthenticationFilter(environment.getProperty("token.secret"));
+        return new StatelessAuthenticationFilter(environment.getProperty("token.secret"), repository);
     }
 }
