@@ -1,13 +1,5 @@
 package outout.unit.services;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.runners.MockitoJUnitRunner;
-import org.springframework.boot.test.SpringApplicationConfiguration;
-import outout.OutoutApplication;
 import outout.model.User;
 import outout.repository.AccountRepository.Exceptions.AccountRepositoryException;
 import outout.repository.AccountRepository.IAccountRepository;
@@ -15,20 +7,28 @@ import outout.services.AccountService.IAccountService;
 import outout.services.AccountService.Impl.AccountService;
 import outout.view.AccountCredentials;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
-@RunWith(MockitoJUnitRunner.class)
-@SpringApplicationConfiguration(classes = OutoutApplication.class)
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
+
+@ExtendWith(MockitoExtension.class)
 public class AccountServiceTests {
 
     @Mock
     private IAccountRepository accountRepository;
 
+    @Mock
+    private PasswordEncoder passwordEncoder;
 
     @InjectMocks
     private IAccountService accountService = new AccountService();
-
-
 
     @Test
     public void Should_Create_Account() throws AccountRepositoryException {
@@ -41,7 +41,6 @@ public class AccountServiceTests {
         accountService.createAccount(accountCredentials);
         //Assert
         Mockito.verify(accountRepository, Mockito.times(1)).createAccount(Mockito.any(User.class));
-
     }
 
     @Test
@@ -76,5 +75,4 @@ public class AccountServiceTests {
             assertEquals(AccountRepositoryException.class, e.getClass());
         }
     }
-
 }
