@@ -1,47 +1,61 @@
 package outout.security;
 
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
-
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import java.util.ArrayList;
 import java.util.Collection;
 
-public class UserAuthentication implements Authentication {
+
+public class UserAuthentication implements UserDetails {
     private String username;
+    private boolean isAuthenticated = true;
+
     public UserAuthentication(String username) {
         this.username = username;
     }
-    @Override
+
+
     public Collection<? extends GrantedAuthority> getAuthorities() {
+        SimpleGrantedAuthority userAuthority = new SimpleGrantedAuthority("ROLE_USER");
+        Collection<GrantedAuthority> authorities = new ArrayList<>();
+        authorities.add(userAuthority);
+        return authorities;
+    }
+
+    @Override
+    public String getPassword() {
         return null;
     }
 
     @Override
-    public Object getCredentials() {
-        return null;
-    }
-
-    @Override
-    public Object getDetails() {
-        return null;
-    }
-
-    @Override
-    public Object getPrincipal() {
+    public String getUsername() {
         return username;
     }
 
     @Override
-    public boolean isAuthenticated() {
+    public boolean isAccountNonExpired() {
         return true;
     }
 
     @Override
-    public void setAuthenticated(final boolean isAuthenticated) throws IllegalArgumentException {
-
+    public boolean isAccountNonLocked() {
+        return true;
     }
 
     @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+
     public String getName() {
         return username;
     }
+
 }
